refactor: changed logger config folder name
