+++ src/WebApi/Config/LoggerSetup/ColumWriters/GuidColumnWriter.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.WebApi.Config.SerilogSetup.ColumWriters;
+﻿namespace IAVH.BioTablero.CM.WebApi.Config.LoggerSetup.ColumWriters;
 
 using System;
 
@@ -8,11 +8,11 @@ using Serilog.Events;
 using Serilog.Sinks.PostgreSQL;
 
 /// <summary>
-/// Custom Serilog column writer for Boolean data
+/// Custom Serilog column writer for GUID data
 /// </summary>
 /// <param name="propertyName">Serilog property name</param>
 /// <param name="dbType">Column database type</param>
-public class BoolColumnWriter(string propertyName, NpgsqlDbType dbType = NpgsqlDbType.Boolean) : ColumnWriterBase(dbType)
+public class GuidColumnWriter(string propertyName, NpgsqlDbType dbType = NpgsqlDbType.Uuid) : ColumnWriterBase(dbType)
 {
     private readonly string propertyName = propertyName;
 
@@ -26,12 +26,12 @@ public class BoolColumnWriter(string propertyName, NpgsqlDbType dbType = NpgsqlD
     {
         if (logEvent != null && logEvent.Properties.TryGetValue(propertyName, out var value))
         {
-            if (value is ScalarValue scalar && scalar.Value is bool guidValue)
+            if (value is ScalarValue scalar && scalar.Value is Guid guidValue)
             {
                 return guidValue;
             }
         }
 
-        throw new InvalidCastException("Property is missing or not a Boolean.");
+        throw new InvalidCastException("Property is missing or not a Guid.");
     }
 }
